+++ src/fr/maxlego08/menu/MenuItemStack.java
@@ -7,6 +7,7 @@ import fr.maxlego08.menu.api.attribute.IAttribute;
 import fr.maxlego08.menu.api.enchantment.Enchantments;
 import fr.maxlego08.menu.api.enchantment.MenuEnchantment;
 import fr.maxlego08.menu.api.enums.MenuItemRarity;
+import fr.maxlego08.menu.api.font.FontImage;
 import fr.maxlego08.menu.api.itemstack.TrimConfiguration;
 import fr.maxlego08.menu.api.loader.MaterialLoader;
 import fr.maxlego08.menu.api.utils.MapConfiguration;
@@ -303,12 +304,13 @@ public class MenuItemStack extends ZUtils {
         Material finalMaterial = itemStack.getType();
         ItemMeta itemMeta = itemStack.getItemMeta();
         String locale = findPlayerLocale(player);
+        FontImage fontImage = this.inventoryManager.getFontImage();
 
         if (itemMeta != null) {
             if (this.displayName != null) {
                 try {
                     String displayName = locale == null ? this.displayName : this.translatedDisplayName.getOrDefault(locale, this.displayName);
-                    Meta.meta.updateDisplayName(itemMeta, placeholders.parse(displayName), offlinePlayer == null ? player : offlinePlayer);
+                    Meta.meta.updateDisplayName(itemMeta, fontImage.replace(placeholders.parse(displayName)), offlinePlayer == null ? player : offlinePlayer);
                 } catch (Exception exception) {
                     Logger.info("Error with update display name for item " + path + " in file " + filePath + " (" + player + ", " + this.displayName + ")", Logger.LogType.ERROR);
                     exception.printStackTrace();
@@ -317,7 +319,8 @@ public class MenuItemStack extends ZUtils {
 
             if (!this.lore.isEmpty()) {
                 List<String> lore = papi(placeholders.parse(locale == null ? this.lore : this.translatedLore.getOrDefault(locale, this.lore)), player, useCache);
-                lore = lore.stream().flatMap(str -> Arrays.stream(str.split("\n"))).collect(Collectors.toList());
+                lore = lore.stream().flatMap(str -> Arrays.stream(str.split("\n"))).map(fontImage::replace).collect(Collectors.toList());
+
                 Meta.meta.updateLore(itemMeta, lore, offlinePlayer == null ? player : offlinePlayer);
             }
 
@@ -371,20 +374,26 @@ public class MenuItemStack extends ZUtils {
     }
 
     private void buildNewItemStackAPI(ItemStack itemStack, ItemMeta itemMeta, Player player, Placeholders placeholders) {
-        if (this.maxStackSize > 0) {
+        if (this.maxStackSize > 0 && this.maxStackSize != itemStack.getMaxStackSize()) {
             itemMeta.setMaxStackSize(this.maxStackSize);
         }
 
         if (itemMeta instanceof Damageable) {
             Damageable damageable = (Damageable) itemMeta;
 
-            if (this.maxDamage > 0) damageable.setMaxDamage(this.maxDamage);
-            if (this.damage != 0) damageable.setDamage(this.damage);
+            if (this.maxDamage > 0) {
+                damageable.setMaxDamage(this.maxDamage);
+            }
+            if (this.damage != 0) {
+                damageable.setDamage(this.damage);
+            }
 
-            if (this.unbreakableEnabled != null) damageable.setUnbreakable(this.unbreakableEnabled);
+            if (this.unbreakableEnabled != null) {
+                damageable.setUnbreakable(this.unbreakableEnabled);
 
-            if (this.unbreakableShowInTooltip != null && !this.unbreakableShowInTooltip) {
-                itemMeta.addItemFlags(ItemFlag.HIDE_UNBREAKABLE);
+                if (this.unbreakableShowInTooltip != null && !this.unbreakableShowInTooltip) {
+                    itemMeta.addItemFlags(ItemFlag.HIDE_UNBREAKABLE);
+                }
             }
         }
 
@@ -392,7 +401,9 @@ public class MenuItemStack extends ZUtils {
             ((Repairable) itemMeta).setRepairCost(this.repairCost);
         }
 
-        if (this.hideTooltip != null) itemMeta.setHideTooltip(this.hideTooltip);
+        if (this.hideTooltip != null) {
+            itemMeta.setHideTooltip(this.hideTooltip);
+        }
         if (this.hideAdditionalTooltip != null && this.hideAdditionalTooltip) {
             for (ItemFlag value : ItemFlag.values()) {
                 itemMeta.addItemFlags(value);
@@ -407,7 +418,9 @@ public class MenuItemStack extends ZUtils {
             itemMeta.setEnchantmentGlintOverride(this.enchantmentGlint);
         }
 
-        if (this.fireResistant != null) itemMeta.setFireResistant(this.fireResistant);
+        if (this.fireResistant != null) {
+            itemMeta.setFireResistant(this.fireResistant);
+        }
 
         if (this.attributeShowInTooltip != null && !this.attributeShowInTooltip) {
             itemMeta.addItemFlags(ItemFlag.HIDE_ATTRIBUTES);