+++ src/fr/maxlego08/menu/command/VCommand.java
@@ -561,7 +561,7 @@ public abstract class VCommand extends Arguments {
         message(this.sender, Message.DOCUMENTATION_INFORMATION);
         this.subVCommands.forEach(command -> {
             if (command.getPermission() == null || this.sender.hasPermission(command.getPermission())) {
-                message(this.sender, Message.COMMAND_SYNTAXE_HELP, "%syntax%", command.getSyntax(), "%description%",
+                message(this.sender, Message.COMMAND_SYNTAX_HELP, "%syntax%", command.getSyntax(), "%description%",
                         command.getDescription());
             }
         });